fix(core): refactored ai brief code
refs. metriport/metriport-internal#2526
